backend/requirements.txt
python-dotenv
psycopg2
SQLAlchemy
Flask